Se corrigieron cosas para subir a heroku

+++ src/controllers/reservation.controller.js
@@ -430,7 +430,12 @@ exports.getReservation = async (req,res) =>
     {
         const reservationID = req.params.id
         const reservation = await Reservation.findOne({_id:reservationID}).populate('hotel room.room');
-        return res.send({reservation})
+        let services = [];
+        for(let serviceId of reservation.services){
+            const service = await Service.findOne({_id: serviceId.service}).populate('hotel');
+            services.push(service)
+        }
+        return res.send({reservation, services})
     }
     catch(err)
     {
@@ -447,29 +452,11 @@ exports.deleteReservation = async (req, res) => {
         const reservationExist = await Reservation.findOne({ _id: reservationID });
         if (!reservationExist) return res.status(400).send({ message: 'Reservation not found or already deleted.' });
 
-        const date = new Date();
-        const dateLocal = (date).toLocaleString('UTC', { timeZone: 'America/Guatemala' });
-        const splitDate = dateLocal.split(' ');
-        const splitDateOne = splitDate[0].split('/');
-        if (splitDateOne[0] < 10) {
-            splitDateOne[0] = '0' + splitDateOne[0];
-        }
-        if (splitDateOne[1] < 10) {
-            splitDateOne[1] = '0' + splitDateOne[1];
-        }
-        const setDate = splitDateOne[2] + '-' + splitDateOne[1] + '-' + splitDateOne[0];
-        const setDateRoom = new Date(setDate);
-
-
         const reservationExisted = await Reservation.find({ _id: reservationID });
-        for (let reservationDeleted of reservationExisted) {
-
-            const idUpdatedRoom = await Room.findByIdAndUpdate({ _id: reservationDeleted.room.room }, { state: false }, { dateAvailable: setDateRoom });
-
-        }
+        if(!reservationExisted) return res.status(400).send({message: 'Error canceling reservation'})
 
         const reservationDeleted = await Reservation.findOneAndDelete({ _id: reservationID });
-        return res.send({ message: 'Delete Reservation.', reservationDeleted });
+        return res.send({ message: 'Your booking cancellation is successfully.', reservationDeleted });
 
     } catch (err) {
         console.log(err);
